Defining behavior for solution.
Textboxes have been validated in such a way that both latitude and longitude values are provided to constitute proper coordinates. Moreover, upon storing the correct coordinates, they are used to search the place on Google Maps. Should it be found, the place will be displayed on the embedded Google Map.